lstm_glove.twitter.27B/requirements.txt
flask==1.1.2
jupyter==1.0.0
keras==2.4.3
matplotlib==3.3.3
nltk==3.5
h5py==2.10.0
numpy==1.18.5
pandas==1.1.4
scikit-learn==0.23.2
scipy==1.4.1
tweepy==3.9.0
tensorflow-cpu==2.3.0
simplejson==3.17.2
requests==2.25.0
wordcloud==1.8.1